Conflicts:
	CodigoJogo/src/implementations/combate/CEngine.java

+++ rpgCyber-core/src/implementations/combate/CEngine.java
@@ -34,7 +34,7 @@ public class CEngine {
 				//if (listaH.get(H.tipo-1) != null) H.pos = H.tipo;
 				//else H.pos = H.tipo - 1;
 
-				listaH.add(H.tipo-1, H);
+				listaH.add((int)H.tipo-1, H);
 				
 				
 				// se a lista esta vazia, adiciona na primeira posicao
@@ -74,7 +74,7 @@ public class CEngine {
 				//if (listaV.get(V.tipo-1) != null) V.pos = V.tipo;
 				//else V.pos = V.tipo - 1;
 
-				listaV.add(V.tipo-1, V);
+				listaV.add((int)V.tipo-1, V);
 				
 				if(listaI.isEmpty()) {
